requirements.txt
# FastAPI 核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# 向量数据库
pgvector==0.2.4

# 时区处理
pytz==2023.3

# 认证和安全
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP 客户端
httpx==0.25.2
aiohttp==3.9.1

# LLM 客户端
openai==1.6.0
anthropic==0.7.8
google-generativeai==0.3.2

# 工具和任务
celery==5.3.4
pyyaml==6.0.1
jsonschema==4.20.0

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.7 